# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Planner API",
        "description": "Event Planner API is a Laravel API designed to manage events. It allows users to create, read, update, and delete events. Additionally, the API provides authentication features, including email verification and Google sign-up.",
        "contact": {
            "name": "Taras Maltsev",
            "url": "https://github.com/TarasIT",
            "email": "taras.maltsev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://event-planner-api.onrender.com/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Gets list of events",
                "operationId": "7fbb26b04c7f15c21f8ee550ba22bcf5",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of events per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Searches term for events",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Searches events by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorts events by title, date or priority",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "date",
                                "priority"
                            ],
                            "example": "date"
                        }
                    },
                    {
                        "name": "ascending",
                        "in": "query",
                        "description": "Sorts events by title, date or priority in ascending or descending order",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "enum": [
                                "true",
                                "false"
                            ],
                            "example": "true"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ResponseEventScheme"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No events found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No events found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to get events. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Event"
                ],
                "summary": "Creates a new event",
                "description": "Creates a new event",
                "operationId": "f5bbf1b691e038d3f5dbf74c508d481d",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateEvent"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ResponseEventScheme"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create an event",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to create an event. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Event"
                ],
                "summary": "Deletes all events",
                "operationId": "2915478ad1f8677f11a66edfd0be937d",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All events deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "All events deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No events found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No events found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete all events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to delete all events. Please, try later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateEvent": {
                "required": [
                    "title",
                    "date",
                    "time"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the event",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the event",
                        "type": "string"
                    },
                    "date": {
                        "description": "Date of the event",
                        "type": "string"
                    },
                    "time": {
                        "description": "Time of the event",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the event",
                        "type": "string"
                    },
                    "category": {
                        "description": "Category of the event",
                        "type": "string"
                    },
                    "priority": {
                        "description": "Priority of the event",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "picture": {
                        "description": "Picture file (jpeg, png, jpg, avif, gif, webp, svg)",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object",
                "example": {
                    "title": "Meeting",
                    "description": "This meeting is very important",
                    "date": "01/07/2024",
                    "time": "09:00 AM",
                    "location": "Kyiv",
                    "category": "Business",
                    "picture": "binary data of the picture",
                    "priority": "high"
                }
            },
...
            "ResponseEventScheme": {
                "properties": {
                    "id": {
                        "description": "Event ID",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the event",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the event",
                        "type": "string"
                    },
                    "date": {
                        "description": "Date of the event",
                        "type": "string"
                    },
                    "time": {
                        "description": "Time of the event",
                        "type": "string"
                    },
                    "location": {
                        "description": "Location of the event",
                        "type": "string"
                    },
                    "category": {
                        "description": "Category of the event",
                        "type": "string"
                    },
                    "priority": {
                        "description": "Priority of the event",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "picture": {
                        "description": "Picture of the event",
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": "4",
                    "title": "Meeting",
                    "description": "This meeting is very important",
                    "date": "01/07/2024",
                    "time": "09:00 AM",
                    "location": "Kyiv",
                    "category": "Business",
                    "picture": "path/to/the/picture/in/the/cloud/storage",
                    "priority": "high"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}